+++ makefile
@@ -2,7 +2,7 @@ build:
 ifeq (,$(wildcard ./dist))
 	mkdir dist
 endif
-	clang ./src/main.c ./src/string.c -o ./dist/string_lib
+	clang ./src/main.c ./src/string_lib.c -o ./dist/string_lib
 
 use:
 ifeq (,$(wildcard ./dist))
