+++ src/main.c
@@ -1,11 +1,20 @@
 #include <stdio.h>
-#include <stdlib.h>
 
-#include "string.h"
+#include "string_lib.h"
 
 char* assert_int(int, int);
 char* assert_true(int);
 char* assert_false(int);
+
+int strcmp(const char* s1, const char* s2) {
+ printf("here \n");
+
+ if (*s1 != 0 && *s2 != 0) {
+  return 1;
+ }
+
+ return 0;
+}
 
 void run_tests(void) {
   printf(" ============= string test =============\n");
@@ -16,7 +25,8 @@ void run_tests(void) {
   printf("- assert null string %s\n", assert_int(strlen(NULL), 0));
 
   printf("-> strcmp:\n");
-  printf("- assert equal string %s\n", assert_true(strcmp("yo", "yo")));
+  strcmp("0", "0");
+  // printf("- assert equal string %s\n", assert_true(strcmp("yo", "yo")));
 }
 
 int main(void) {
