+++ src/string_lib.c
@@ -1,0 +1,17 @@
+#include <stdio.h>
+
+size_t strlen(const char *s) {
+  if (NULL == s) {
+    return 0;
+  }
+
+  size_t length = 0;
+  while (*s != 0) {
+    length++;
+    s++;
+  }
+
+  return length;
+}
+
+
