+++ src/string_lib.h
@@ -4,6 +4,6 @@
 # include <stdlib.h>
 
 size_t strlen(const char *s);
-int strcmp(const char *s1, const char *s2);
+// int strcmp(const char *s1, const char *s2);
 
 #endif /* !STRLEN_H */
